# js/apps/admin/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Admin API",
    "version": "1.0.0",
    "description": "Design system management API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "additionalProperties": {
                  "nullable": true
                }
              }
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/queries": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List queries from the catalog",
        "responses": {
          "200": {
            "description": "Queries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "queries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "params": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string"
                                },
                                "type": {
                                  "type": "string"
                                },
                                "default": {
                                  "nullable": true
                                },
                                "options": {
                                  "type": "array",
                                  "items": {
                                    "nullable": true
                                  }
                                }
                              },
                              "required": [
                                "name",
                                "type"
                              ]
                            }
                          }
                        },
                        "required": [
                          "id",
                          "label",
                          "type"
                        ]
                      }
                    }
                  },
                  "required": [
                    "queries"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/queries/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Run a catalog query",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "result": {
                      "nullable": true
                    },
                    "count": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "label",
                    "count"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}